+++ src/AccountLogin/Exception/Oauth2LoginException.php
@@ -20,7 +20,7 @@
 
 namespace PrestaShop\Module\PsAccounts\AccountLogin\Exception;
 
-use PrestaShop\Module\PsAccounts\OAuth2\Response\UserInfo;
+use PrestaShop\Module\PsAccounts\OAuth2\Resource\UserInfo;
 
 class Oauth2LoginException extends AccountLoginException
 {
