+++ src/AccountLogin/OAuth2Session.php
@@ -23,8 +23,8 @@ namespace PrestaShop\Module\PsAccounts\AccountLogin;
 use PrestaShop\Module\PsAccounts\OAuth2\ApiClient;
 use PrestaShop\Module\PsAccounts\OAuth2\Client;
 use PrestaShop\Module\PsAccounts\OAuth2\OAuth2Exception;
-use PrestaShop\Module\PsAccounts\OAuth2\Response\AccessToken;
-use PrestaShop\Module\PsAccounts\OAuth2\Response\UserInfo;
+use PrestaShop\Module\PsAccounts\OAuth2\Resource\AccessToken;
+use PrestaShop\Module\PsAccounts\OAuth2\Resource\UserInfo;
 use PrestaShop\Module\PsAccounts\OAuth2\Token\Validator\Validator;
 use Symfony\Component\HttpFoundation\Session\SessionInterface;
