+++ src/Http/Exception/BadRequestException.php
@@ -18,15 +18,19 @@
  * @license   https://opensource.org/licenses/AFL-3.0 Academic Free License version 3.0
  */
 
-namespace PrestaShop\Module\PsAccounts\Api\Controller;
+namespace PrestaShop\Module\PsAccounts\Http\Exception;
 
-use PrestaShop\Module\PsAccounts\Type\Enum;
-
-class RestMethod extends Enum
+class BadRequestException extends HttpException
 {
-    const INDEX = 'index';
-    const STORE = 'store';
-    const UPDATE = 'update';
-    const DELETE = 'delete';
-    const SHOW = 'show';
+    /**
+     * @param string $message
+     * @param int $code
+     * @param \Exception|null $previous
+     */
+    public function __construct($message = 'Bad Request', $code = 0, \Exception $previous = null)
+    {
+        parent::__construct($message, $code, $previous);
+
+        $this->statusCode = 400;
+    }
 }
