+++ src/Http/Exception/MethodNotAllowedException.php
@@ -18,7 +18,7 @@
  * @license   https://opensource.org/licenses/AFL-3.0 Academic Free License version 3.0
  */
 
-namespace PrestaShop\Module\PsAccounts\Api\Controller\Exception;
+namespace PrestaShop\Module\PsAccounts\Http\Exception;
 
 class MethodNotAllowedException extends HttpException
 {
