+++ src/Http/Request/ShopHealthCheckRequest.php
@@ -18,7 +18,7 @@
  * @license   https://opensource.org/licenses/AFL-3.0 Academic Free License version 3.0
  */
 
-namespace PrestaShop\Module\PsAccounts\Api\Controller\Request;
+namespace PrestaShop\Module\PsAccounts\Http\Request;
 
 class ShopHealthCheckRequest extends Request
 {
