+++ src/OAuth2/Resource/WellKnown.php
@@ -18,9 +18,9 @@
  * @license   https://opensource.org/licenses/AFL-3.0 Academic Free License version 3.0
  */
 
-namespace PrestaShop\Module\PsAccounts\OAuth2\Response;
+namespace PrestaShop\Module\PsAccounts\OAuth2\Resource;
 
-use PrestaShop\Module\PsAccounts\OAuth2\Response\Response;
+use PrestaShop\Module\PsAccounts\Http\Resource\Resource;
 
 /**
  * {
@@ -54,7 +54,7 @@ use PrestaShop\Module\PsAccounts\OAuth2\Response\Response;
  *  "code_challenge_methods_supported":["plain","S256"]
  * }
  */
-class WellKnown extends Response
+class WellKnown extends Resource
 {
     /** @var string */
     public $issuer;
